refactor: ticket 패키지 명칭을 booking으로 변경
- ticket 패키지 내의 모든 클래스를 booking 패키지로 이동
- 관련 클래스들의 패키지 선언 및 임포트 경로 업데이트
- 테스트 코드의 패키지 구조 및 참조 수정
