+++ ticketing/src/main/java/org/truve/platform/ticketing/service/booking/dto/TicketRequest.java
@@ -1,4 +1,4 @@
-package org.truve.platform.ticketing.service.ticket.dto;
+package org.truve.platform.ticketing.service.booking.dto;
 
 import java.util.List;
