+++ ticketing/src/main/java/org/truve/platform/ticketing/service/booking/repository/ReservationRepository.java
@@ -1,7 +1,7 @@
-package org.truve.platform.ticketing.service.ticket.repository;
+package org.truve.platform.ticketing.service.booking.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.truve.platform.ticketing.service.ticket.domain.entity.Reservation;
+import org.truve.platform.ticketing.service.booking.domain.entity.Reservation;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 }
